+++ pymongo/asynchronous/encryption.py
@@ -463,7 +463,14 @@ class Algorithm(str, enum.Enum):
     RANGE = "Range"
     """Range.
 
-    .. versionadded:: 4.8
+    .. versionadded:: 4.9
+    """
+    RANGEPREVIEW = "RangePreview"
+    """**DEPRECATED** - RangePreview.
+
+    .. note:: Support for RangePreview is deprecated. Use :attr:`Algorithm.RANGE` instead.
+
+    .. versionadded:: 4.4
     """
 
 
@@ -477,7 +484,18 @@ class QueryType(str, enum.Enum):
     """Used to encrypt a value for an equality query."""
 
     RANGE = "range"
-    """Used to encrypt a value for a range query."""
+    """Used to encrypt a value for a range query.
+
+    .. versionadded:: 4.9
+    """
+
+    RANGEPREVIEW = "RangePreview"
+    """**DEPRECATED** - Used to encrypt a value for a rangePreview query.
+
+    .. note:: Support for RangePreview is deprecated. Use :attr:`QueryType.RANGE` instead.
+
+    .. versionadded:: 4.4
+    """
 
 
 def _create_mongocrypt_options(**kwargs: Any) -> MongoCryptOptions:
@@ -847,7 +865,7 @@ class ClientEncryption(Generic[_DocumentType]):
 
         :return: The encrypted value, a :class:`~bson.binary.Binary` with subtype 6.
 
-        .. versionchanged:: 4.8
+        .. versionchanged:: 4.9
            Added the `range_opts` parameter.
 
         .. versionchanged:: 4.7
@@ -903,7 +921,7 @@ class ClientEncryption(Generic[_DocumentType]):
 
         :return: The encrypted expression, a :class:`~bson.RawBSONDocument`.
 
-        .. versionchanged:: 4.8
+        .. versionchanged:: 4.9
            Added the `range_opts` parameter.
 
         .. versionchanged:: 4.7